Respect line ending when copy selected text to clipboard
Respect line ending when copy selected text to clipboard
